+++ examples/integration_hybrid_100s.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import numpy as np
 
 import pykilosort
-from deploy.serverpc.kilosort2.run_pykilosort import run_spike_sorting_ibl
+from pykilosort.ibl import run_spike_sorting_ibl
 
 # Minimum recall to pass checks (percentage)
 MIN_RECALL = 90