+++ pykilosort/ibl.py
@@ -10,6 +10,7 @@ import spikeglx
 import neuropixel
 from ibllib.ephys import spikes
 from one.alf.files import get_session_path
+from one.remote import aws
 from pykilosort import add_default_handler, run, Bunch, __version__
 from pykilosort.params import KilosortParams
 
@@ -90,6 +91,9 @@ def run_spike_sorting_ibl(bin_file, scratch_dir=None, delete=True,
         _logger.info(f"Starting Pykilosort version {__version__}")
         _logger.info(f"Scratch dir {ks_output_dir}")
         _logger.info(f"Output dir {bin_file.parent}")
+        _logger.info(f"Log file {log_file}")
+        _logger.info(f"Loaded probe geometry for NP{params['probe']['neuropixel_version']}")
+
         run(bin_file, dir_path=scratch_dir, output_dir=ks_output_dir, **params)
         # move back the QC files to the original probe folder
         for qc_file in scratch_dir.joinpath(".kilosort", bin_file.name).glob('_iblqc_*'):
@@ -114,6 +118,7 @@ def ibl_pykilosort_params(bin_file):
     params = KilosortParams()
     params.preprocessing_function = 'destriping'
     params.probe = probe_geometry(bin_file)
+    params.minFR = 0
     # params = {k: dict(params)[k] for k in sorted(dict(params))}
     return dict(params)
 
@@ -140,12 +145,21 @@ def probe_geometry(bin_file):
     probe = Bunch()
     probe.NchanTOT = nc + 1
     probe.chanMap = np.arange(nc)
-    probe.xc = h['x']
+    probe.xc = h['x'] + h['shank'] * 200
     probe.yc = h['y']
     probe.x = h['x']
     probe.y = h['y']
+    probe.shank = h['shank']
     probe.kcoords = np.zeros(nc)
     probe.neuropixel_version = ver
     probe.sample_shift = h['sample_shift']
     probe.h = h
     return probe
+
+
+def download_test_data(local_folder):
+    return aws.s3_download_folder('spikesorting/integration_tests', local_folder)
+
+
+def download_benchmark_data(local_folder):
+    return aws.s3_download_folder('spikesorting/benchmarks', local_folder)